feat: error_handling wall and duplicate ==== Prefix ==== - feat: 新しい機能 - fix: バグの修正 - docs: ドキュメントのみの変更 - style: 空白、フォーマット、セミコロン追加など - refactor: 仕様に影響がないコード改善(リファクタ) - perf: パフォーマンス向上関連 - test: テスト関連 - chore: ビルド、補助ツール、ライブラリ関連 - wip: 一時中断

+++ lib/printf/libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: yehara <yehara@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/23 11:40:23 by yehara            #+#    #+#             */
-/*   Updated: 2024/05/08 21:02:50 by yehara           ###   ########.fr       */
+/*   Updated: 2024/09/16 19:36:56 by ebarayuug        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int ft_strlen(const char *s)
 {
 	size_t	i;
